fix query
