+++ database/seeders/TagsTableSeeder.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use App\Models\Tag;
 use Faker\Factory as Faker;
-use database\factories\TagFactory;
+use Database\Factories\TagFactory;
 
 class TagsTableSeeder extends Seeder
 {
