+++ resources/views/courses/list_course.blade.php
