+++ resources/views/layouts/header.blade.php
@@ -10,7 +10,7 @@
             <div class="collapse navbar-collapse ml-auto" id="navbarResponsive">
                 <ul class="navbar-nav ml-auto">
                     <li class="nav-item"><a class="nav-link nav-link-active active {{ Route::currentRouteName() == 'home' ? 'active' : '' }}" href="{{route('home')}}">HOME</a></li>
-                    <li class="nav-item"><a class="nav-link {{ Route::currentRouteName() == 'all-courses' ? 'active' : '' }}" href="{{ route('all-courses') }}">ALL COURSE</a></li>
+                    <li class="nav-item"><a class="nav-link {{ Route::currentRouteName() == 'courses' ? 'active' : '' }}" href="{{ route('courses') }}">ALL COURSE</a></li>
                     @guest
                         @if (Route::has('login') || Route::has('register'))
                         <li class="nav-item"><a class="nav-link header-nav" href="{{ route('login') }}" data-toggle="modal" data-target="#loginRegisterModal">LOGIN/REGISTER</a></li>
